fixed the home page so it renders product pictures from the new backend routs
